tudo eh float; ALIEN se move ateh o ponto, mas buga aas vezes; parar de tremer o alien ao chegar ao ponto

+++ include/Action.h
@@ -7,7 +7,7 @@ class Action {
         enum ActionType {MOVE=0, SHOOT};
         ActionType type;
         Vec2 pos;
-        Action(ActionType, float, float);
+        Action(ActionType, double, double);
 };
