+++ src/Action.cpp
@@ -1,6 +1,6 @@
 #include "Action.h"
 
-Action::Action(ActionType actionType, float x, float y) {
+Action::Action(ActionType actionType, double x, double y) {
     this->type = actionType;
     this->pos = {x, y};
 }
