+++ src/Vec2.cpp
@@ -2,12 +2,12 @@
 #include <cmath>
 #define PI M_PI
 
-// static float lineSlope(const Vec2& va, const Vec2& vb) {
+// static double lineSlope(const Vec2& va, const Vec2& vb) {
 //     return ( (va.x - vb.x)/(va.y - vb.y) );
 // }
 
 Vec2::Vec2() : x(0), y(0){}
-Vec2::Vec2(float x, float y) : x(x), y(y){}
+Vec2::Vec2(double x, double y) : x(x), y(y){}
 
 // Operações elementares entre vetores
 
@@ -40,7 +40,7 @@ Vec2 Vec2::operator-(const int& scalar) const{
     return {this->x  - scalar, this->y  - scalar};
 }
 
-float Vec2::abs() {
+double Vec2::abs() {
     return sqrt(this->x*this->x + this->y*this->y);
 }
 
@@ -49,25 +49,25 @@ Vec2 Vec2::unitary() {
     return Vec2(this->x/this->norm, this->y/this->norm);
 }
 
-float Vec2::dist(const Vec2& point) const {
-    float deltaX = this->x-point.x;
-    float deltaY = this->x-point.y;
+double Vec2::dist(const Vec2& point) const {
+    double deltaX = this->x-point.x;
+    double deltaY = this->x-point.y;
     return sqrt( deltaX*deltaX + deltaY*deltaY );
 }
 
-float Vec2::xAxisSlope() const{
+double Vec2::xAxisSlope() const{
     return atan2(this->y, this->x);
 }
 
-void Vec2::rotate(const float& degree) {
-    const float radian = degree * DEG2RAD;
+void Vec2::rotate(const double& degree) {
+    const double radian = degree * DEG2RAD;
     auto xOld = this->x;
     auto yOld = this->y;
     this->x = xOld * cos(radian) - yOld * sin(radian);
     this->y = yOld * cos(radian) + xOld * sin(radian);
 }
 
-Vec2 Vec2::GetRotated(const float& degree) {
+Vec2 Vec2::GetRotated(const double& degree) {
     Vec2 toRotate(this->x, this->y);
     toRotate.rotate(degree);
     return toRotate;
